Change SalesManager

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {1, 44, 788, 90};
+        long[] sales = {1, 44, 788, 90};
         SalesManager manager = new SalesManager(sales);
         System.out.println("Макс  " + manager.max());
         }
